requirements.txt
fastapi
uvicorn[standard]
python-multipart
httpx
supabase
pandas
weaviate-client
langchain
langchain-community
langchain-weaviate
langchain-groq
sentence-transformers
faiss-cpu
python-dotenv
pydantic
email-validator
groq